requirements.txt
Django==4.1.3
djangorestframework
environs
psycopg2-binary 
python-decouple
djangorestframework-simplejwt
django-cors-headers